+++ src/app/movie/[id]/[category]/components/ReviewsTab.tsx
@@ -1,14 +1,12 @@
 'use client';
 
-import { deleteReview, review, ReviewState, updateReview } from '@/app/movie/lib/actions';
-import { MovieDetailResponseDto, Review } from '@/types/movie-detail-response-dto';
+import { MovieDetailResponseDto } from '@/types/movie-detail-response-dto';
 import { UserWithMovieIds } from '@/types/user';
 import { fetchMovieReviews } from '@/utils/api';
-import { useVirtualizer } from '@tanstack/react-virtual';
-import Link from 'next/link';
-import { useActionState, useRef, useState } from 'react';
-import commonStyles from './styles/common.module.css';
+import { useRef, useState } from 'react';
 import styles from './styles/reviews.module.css';
+import ReviewForm from './reviews/ReviewForm';
+import ReviewList from './reviews/ReviewList';
 
 interface ReviewsTabProps {
   movie: MovieDetailResponseDto;
@@ -16,71 +14,12 @@ interface ReviewsTabProps {
   category: string;
 }
 
-interface ReviewItemProps {
-  review: Review;
-  currentUserId?: string;
-  style: React.CSSProperties;
-  movie: MovieDetailResponseDto;
-  category: string;
-  onRefresh?: () => void;
-}
-
-const ReviewItem = ({ review, currentUserId, style, movie, category, onRefresh }: ReviewItemProps) => {
-  const isAuthor = currentUserId === review.profile?.id;
-  const [deleteState, deleteAction] = useActionState(deleteReview, { error: '', message: '' });
-
-  const renderActionError = (error: string | Record<string, string[]>) => {
-    if (typeof error === 'string') {
-      return <p className={styles.error}>{error}</p>;
-    }
-    return null;
-  };
-
-  return (
-    <div className={styles.reviewItem} style={style}>
-      <div className={styles.reviewContent}>
-        <div className={styles.reviewHeader}>
-          <div className={styles.authorInfo}>
-            <div className={styles.authorName}>{review.profile?.name || '알 수 없음'}</div>
-          </div>
-          {isAuthor && (
-            <div className={styles.reviewActions}>
-              <form action={deleteAction} style={{ display: 'inline' }}>
-                <input type="hidden" name="reviewId" value={review.id} />
-                <input type="hidden" name="userId" value={review.profile?.id} />
-                <button 
-                  type="submit" 
-                  className={styles.deleteButton}
-                  onClick={(e) => {
-                    if (!window.confirm('정말로 이 리뷰를 삭제하시겠습니까?')) {
-                      e.preventDefault();
-                    }
-                  }}
-                >
-                  삭제
-                </button>
-              </form>
-            </div>
-          )}
-        </div>
-        {deleteState.error && renderActionError(deleteState.error)}
-        <p className={styles.reviewText}>{review.content}</p>
-      </div>
-    </div>
-  );
-};
-
 export default function ReviewsTab({ movie, userWithMovieIds, category }: ReviewsTabProps) {
-  const initialState: ReviewState = {
-    error: {},
-    message: ""
-  }
   const { user, review_movieIds } = userWithMovieIds;
   const isLogin = user !== null;
   const isReviewed = review_movieIds.includes(Number(movie.theMovieDbId));
-  const [reviewState, reviewAction] = useActionState(review, initialState);
   const [currentPage, setCurrentPage] = useState(1);
-  const [reviews, setReviews] = useState<Review[]>(movie.recentReviews || []);
+  const [reviews, setReviews] = useState(movie.recentReviews || []);
   const parentRef = useRef<HTMLDivElement>(null);
   const hasReviews = movie.totalReviews > 0;
 
@@ -94,109 +33,28 @@ export default function ReviewsTab({ movie, userWithMovieIds, category }: Review
     }
   };
 
-  const rowVirtualizer = useVirtualizer({
-    count: reviews.length,
-    getScrollElement: () => parentRef.current,
-    estimateSize: () => 100,
-    overscan: 5,
-  });
-
-  const renderError = () => {
-    if (typeof reviewState.error === 'string') {
-      return <p className={commonStyles.error}>{reviewState.error}</p>;
-    } else if (reviewState.error) {
-      return (
-        <ul className={styles.errorList}>
-          {Object.entries(reviewState.error).map(([key, errors]) => (
-            errors && errors.map((error, index) => (
-              <li key={`${key}-${index}`} className={styles.errorItem}>{error}</li>
-            ))
-          ))}
-        </ul>
-      );
-    }
-    return null;
-  };
-
   return (
     <div className={styles.review}>
-      {isLogin && !isReviewed ? (
-        <div className={styles.reviewForm}>
-          <h3 className={styles.reviewTitle}>리뷰 작성하기</h3>
-          <form action={reviewAction} role='form'>
-            <textarea
-              name='review'
-              placeholder="이 영화에 대한 리뷰를 작성해주세요..."
-              className={styles.reviewInput}
-              required
-            />
-            <input type="hidden" name="movie_id" value={movie.id} />
-            <input type="hidden" name="user_id" value={user.id} />
-            <input type="hidden" name="the_movie_db_id" value={movie.theMovieDbId} />
-            <input type="hidden" name="category" value={category} />
-            <button type="submit">리뷰 제출</button>
-          </form>
-          {renderError()}
-          {reviewState.message && <p className={styles.message}>{reviewState.message}</p>}
-        </div>
-      ) : isReviewed ? (
-        <p className={styles.rated}>이미 리뷰를 작성했습니다.</p>
-      ) : (
-        <p className={styles.loginPrompt}>
-          <Link href="/login">로그인</Link>하고 리뷰를 작성해보세요!
-        </p>
-      )}
+      <ReviewForm
+        isLogin={isLogin}
+        isReviewed={isReviewed}
+        movieId={movie.id}
+        userId={user?.id || ''}
+        theMovieDbId={movie.theMovieDbId}
+        category={category}
+      />
 
       {hasReviews ? (
-        <>
-          <div ref={parentRef} className={styles.reviewList} style={{ height: '400px', overflow: 'auto' }}>
-            <div
-              style={{
-                height: `${rowVirtualizer.getTotalSize()}px`,
-                width: '100%',
-                position: 'relative',
-              }}
-            >
-              {rowVirtualizer.getVirtualItems().map((virtualRow) => {
-                const review = reviews[virtualRow.index];
-                return (
-                  <ReviewItem
-                    key={virtualRow.index}
-                    review={review}
-                    currentUserId={user?.id}
-                    movie={movie}
-                    category={category}
-                    style={{
-                      position: 'absolute',
-                      top: 0,
-                      left: 0,
-                      width: '100%',
-                      height: `${virtualRow.size}px`,
-                      transform: `translateY(${virtualRow.start}px)`,
-                    }}
-                    onRefresh={() => handlePageChange(currentPage)}
-                  />
-                );
-              })}
-            </div>
-          </div>
-
-          <div className={commonStyles.pagination}>
-            <button
-              onClick={() => handlePageChange(currentPage - 1)}
-              disabled={currentPage === 1}
-            >
-              이전
-            </button>
-            <span>{currentPage}</span>
-            <button
-              onClick={() => handlePageChange(currentPage + 1)}
-              disabled={currentPage * 10 >= movie.totalReviews}
-            >
-              다음
-            </button>
-          </div>
-        </>
+        <ReviewList
+          reviews={reviews}
+          currentUserId={user?.id}
+          movie={movie}
+          category={category}
+          parentRef={parentRef}
+          currentPage={currentPage}
+          totalReviews={movie.totalReviews}
+          onPageChange={handlePageChange}
+        />
       ) : (
           <div className={styles.noReviews}>
             <p>아직 작성된 리뷰가 없습니다.</p>
